#477 企业号接口重构，分类管理
