+++ weixin-java-cp/src/main/java/me/chanjar/weixin/cp/api/WxCpService.java
@@ -19,9 +19,9 @@ import java.util.List;
 
 /**
  * 微信API的Service
+ * @author chanjaster
  */
 public interface WxCpService {
-
   /**
    * <pre>
    * 验证推送过来的消息的正确性
@@ -86,42 +86,6 @@ public interface WxCpService {
   WxJsapiSignature createJsapiSignature(String url) throws WxErrorException;
 
   /**
-   * @deprecated  请使用 {@link WxCpMenuService#create(WxMenu)}
-   */
-  @Deprecated
-  void menuCreate(WxMenu menu) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpMenuService#create(Integer, WxMenu)}
-   */
-  @Deprecated
-  void menuCreate(Integer agentId, WxMenu menu) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpMenuService#delete()}  }
-   */
-  @Deprecated
-  void menuDelete() throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpMenuService#delete(Integer)}
-   */
-  @Deprecated
-  void menuDelete(Integer agentId) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpMenuService#get() }
-   */
-  @Deprecated
-  WxMenu menuGet() throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpMenuService#get(Integer)}
-   */
-  @Deprecated
-  WxMenu menuGet(Integer agentId) throws WxErrorException;
-
-  /**
    * <pre>
    * 发送消息
    * 详情请见: http://qydev.weixin.qq.com/wiki/index.php?title=%E5%8F%91%E9%80%81%E6%8E%A5%E5%8F%A3%E8%AF%B4%E6%98%8E
@@ -130,175 +94,6 @@ public interface WxCpService {
    * @param message 要发送的消息对象
    */
   WxCpMessageSendResult messageSend(WxCpMessage message) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpDepartmentService#create(WxCpDepart)}
-   */
-  @Deprecated
-  Integer departCreate(WxCpDepart depart) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpDepartmentService#update(WxCpDepart)}
-   */
-  @Deprecated
-  void departUpdate(WxCpDepart group) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpDepartmentService#delete(Integer)}
-   */
-  @Deprecated
-  void departDelete(Integer departId) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpDepartmentService#listAll() }
-   */
-  @Deprecated
-  List<WxCpDepart> departGet() throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpMediaService#upload(String, String, InputStream)}
-   */
-  @Deprecated
-  WxMediaUploadResult mediaUpload(String mediaType, String fileType, InputStream inputStream)
-    throws WxErrorException, IOException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpMediaService#upload(String, File)}
-   */
-  @Deprecated
-  WxMediaUploadResult mediaUpload(String mediaType, File file) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpMediaService#download(String)}
-   */
-  @Deprecated
-  File mediaDownload(String mediaId) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpUserService#authenticate(String)}
-   */
-  @Deprecated
-  void userAuthenticated(String userId) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpUserService#create(WxCpUser)}
-   */
-  @Deprecated
-  void userCreate(WxCpUser user) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpUserService#update(WxCpUser)}
-   */
-  @Deprecated
-  void userUpdate(WxCpUser user) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpUserService#delete(String...)}
-   */
-  @Deprecated
-  void userDelete(String userid) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpUserService#delete(String...)}
-   */
-  @Deprecated
-  void userDelete(String[] userids) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpUserService#getById(String)}
-   */
-  @Deprecated
-  WxCpUser userGet(String userid) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpUserService#listByDepartment(Integer, Boolean, Integer)}
-   */
-  @Deprecated
-  List<WxCpUser> userList(Integer departId, Boolean fetchChild, Integer status) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpUserService#listSimpleByDepartment(Integer, Boolean, Integer)}
-   */
-  @Deprecated
-  List<WxCpUser> departGetUsers(Integer departId, Boolean fetchChild, Integer status) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpTagService#create(String)}
-   */
-  @Deprecated
-  String tagCreate(String tagName) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpTagService#update(String, String)}
-   */
-  @Deprecated
-  void tagUpdate(String tagId, String tagName) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpTagService#delete(String)}
-   */
-  @Deprecated
-  void tagDelete(String tagId) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpTagService#listAll()}
-   */
-  @Deprecated
-  List<WxCpTag> tagGet() throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpTagService#listUsersByTagId(String)}
-   */
-  @Deprecated
-  List<WxCpUser> tagGetUsers(String tagId) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpTagService#addUsers2Tag(String, List, List)}
-   */
-  @Deprecated
-  void tagAddUsers(String tagId, List<String> userIds, List<String> partyIds) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpTagService#removeUsersFromTag(String, List)}
-   */
-  @Deprecated
-  void tagRemoveUsers(String tagId, List<String> userIds) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpOAuth2Service#buildAuthorizationUrl(String)}
-   */
-  @Deprecated
-  String oauth2buildAuthorizationUrl(String state);
-
-  /**
-   * @deprecated  请使用 {@link WxCpOAuth2Service#buildAuthorizationUrl(String, String)}
-   */
-  @Deprecated
-  String oauth2buildAuthorizationUrl(String redirectUri, String state);
-
-  /**
-   * @deprecated  请使用 {@link WxCpOAuth2Service#getUserInfo(String)}
-   */
-  @Deprecated
-  String[] oauth2getUserInfo(String code) throws WxErrorException;
-
-  /**
-   * @deprecated  请使用 {@link WxCpOAuth2Service#getUserInfo(Integer, String)}
-   */
-  @Deprecated
-  String[] oauth2getUserInfo(Integer agentId, String code) throws WxErrorException;
-
-  /**
-   * <pre>
-   * 邀请成员关注
-   * http://qydev.weixin.qq.com/wiki/index.php?title=管理成员#.E9.82.80.E8.AF.B7.E6.88.90.E5.91.98.E5.85.B3.E6.B3.A8
-   * </pre>
-   *
-   * @param userId     用户的userid
-   * @param inviteTips 推送到微信上的提示语（只有认证号可以使用）。当使用微信推送时，该字段默认为“请关注XXX企业号”，邮件邀请时，该字段无效。
-   * @return 1:微信邀请 2.邮件邀请
-   */
-  int invite(String userId, String inviteTips) throws WxErrorException;
 
   /**
    * <pre>
