+++ weixin-java-cp/src/main/java/me/chanjar/weixin/cp/api/impl/WxCpServiceAbstractImpl.java
@@ -139,188 +139,6 @@ public abstract class WxCpServiceAbstractImpl<H, P> implements WxCpService, Requ
   }
 
   @Override
-  @Deprecated
-  public void menuCreate(WxMenu menu) throws WxErrorException {
-    this.getMenuService().create(menu);
-  }
-
-  @Override
-  @Deprecated
-  public void menuCreate(Integer agentId, WxMenu menu) throws WxErrorException {
-    this.getMenuService().create(agentId, menu);
-  }
-
-  @Override
-  @Deprecated
-  public void menuDelete() throws WxErrorException {
-    this.getMenuService().delete();
-  }
-
-  @Override
-  @Deprecated
-  public void menuDelete(Integer agentId) throws WxErrorException {
-    this.getMenuService().delete(agentId);
-  }
-
-  @Override
-  @Deprecated
-  public WxMenu menuGet() throws WxErrorException {
-    return this.getMenuService().get();
-  }
-
-  @Override
-  @Deprecated
-  public WxMenu menuGet(Integer agentId) throws WxErrorException {
-    return this.getMenuService().get(agentId);
-  }
-
-  @Override
-  @Deprecated
-  public WxMediaUploadResult mediaUpload(String mediaType, String fileType, InputStream inputStream)
-    throws WxErrorException, IOException {
-    return this.getMediaService().upload(mediaType, fileType, inputStream);
-  }
-
-  @Override
-  @Deprecated
-  public WxMediaUploadResult mediaUpload(String mediaType, File file) throws WxErrorException {
-    return this.getMediaService().upload(mediaType, file);
-  }
-
-  @Override
-  @Deprecated
-  public File mediaDownload(String mediaId) throws WxErrorException {
-    return this.getMediaService().download(mediaId);
-  }
-
-  @Override
-  @Deprecated
-  public void userAuthenticated(String userId) throws WxErrorException {
-    this.getUserService().authenticate(userId);
-  }
-
-  @Override
-  @Deprecated
-  public void userCreate(WxCpUser user) throws WxErrorException {
-    this.getUserService().create(user);
-  }
-
-  @Override
-  @Deprecated
-  public void userUpdate(WxCpUser user) throws WxErrorException {
-    this.getUserService().update(user);
-  }
-
-  @Override
-  @Deprecated
-  public void userDelete(String userid) throws WxErrorException {
-    this.getUserService().delete(userid);
-  }
-
-  @Override
-  @Deprecated
-  public void userDelete(String[] userids) throws WxErrorException {
-    this.getUserService().delete(userids);
-  }
-
-  @Override
-  @Deprecated
-  public WxCpUser userGet(String userid) throws WxErrorException {
-    return this.getUserService().getById(userid);
-  }
-
-  @Override
-  @Deprecated
-  public List<WxCpUser> userList(Integer departId, Boolean fetchChild, Integer status) throws WxErrorException {
-    return this.getUserService().listByDepartment(departId, fetchChild, status);
-  }
-
-  @Override
-  @Deprecated
-  public List<WxCpUser> departGetUsers(Integer departId, Boolean fetchChild, Integer status) throws WxErrorException {
-    return this.getUserService().listSimpleByDepartment(departId, fetchChild, status);
-  }
-
-  @Override
-  @Deprecated
-  public String tagCreate(String tagName) throws WxErrorException {
-    return this.getTagService().create(tagName);
-  }
-
-  @Override
-  @Deprecated
-  public void tagUpdate(String tagId, String tagName) throws WxErrorException {
-    this.getTagService().update(tagId, tagName);
-  }
-
-  @Override
-  @Deprecated
-  public void tagDelete(String tagId) throws WxErrorException {
-    this.getTagService().delete(tagId);
-  }
-
-  @Override
-  @Deprecated
-  public List<WxCpTag> tagGet() throws WxErrorException {
-    return this.getTagService().listAll();
-  }
-
-  @Override
-  @Deprecated
-  public List<WxCpUser> tagGetUsers(String tagId) throws WxErrorException {
-    return this.getTagService().listUsersByTagId(tagId);
-  }
-
-  @Override
-  @Deprecated
-  public void tagAddUsers(String tagId, List<String> userIds, List<String> partyIds) throws WxErrorException {
-    this.getTagService().addUsers2Tag(tagId, userIds, partyIds);
-  }
-
-  @Override
-  @Deprecated
-  public void tagRemoveUsers(String tagId, List<String> userIds) throws WxErrorException {
-    this.getTagService().removeUsersFromTag(tagId, userIds);
-  }
-
-  @Override
-  @Deprecated
-  public String oauth2buildAuthorizationUrl(String state) {
-    return this.getOauth2Service().buildAuthorizationUrl(state);
-  }
-
-  @Override
-  @Deprecated
-  public String oauth2buildAuthorizationUrl(String redirectUri, String state) {
-    return this.getOauth2Service().buildAuthorizationUrl(redirectUri, state);
-  }
-
-  @Override
-  @Deprecated
-  public String[] oauth2getUserInfo(String code) throws WxErrorException {
-    return this.getOauth2Service().getUserInfo(code);
-  }
-
-  @Override
-  @Deprecated
-  public String[] oauth2getUserInfo(Integer agentId, String code) throws WxErrorException {
-    return this.getOauth2Service().getUserInfo(agentId, code);
-  }
-
-  @Override
-  public int invite(String userId, String inviteTips) throws WxErrorException {
-    String url = "https://qyapi.weixin.qq.com/cgi-bin/invite/send";
-    JsonObject jsonObject = new JsonObject();
-    jsonObject.addProperty("userid", userId);
-    if (StringUtils.isNotEmpty(inviteTips)) {
-      jsonObject.addProperty("invite_tips", inviteTips);
-    }
-    String responseContent = post(url, jsonObject.toString());
-    JsonElement tmpJsonElement = new JsonParser().parse(responseContent);
-    return tmpJsonElement.getAsJsonObject().get("type").getAsInt();
-  }
-
-  @Override
   public String[] getCallbackIp() throws WxErrorException {
     String url = "https://qyapi.weixin.qq.com/cgi-bin/getcallbackip";
     String responseContent = get(url, null);
@@ -514,30 +332,6 @@ public abstract class WxCpServiceAbstractImpl<H, P> implements WxCpService, Requ
   @Override
   public WxCpUserService getUserService() {
     return userService;
-  }
-
-  @Override
-  @Deprecated
-  public Integer departCreate(WxCpDepart depart) throws WxErrorException {
-    return this.getDepartmentService().create(depart);
-  }
-
-  @Override
-  @Deprecated
-  public void departUpdate(WxCpDepart depart) throws WxErrorException {
-    this.getDepartmentService().update(depart);
-  }
-
-  @Override
-  @Deprecated
-  public void departDelete(Integer departId) throws WxErrorException {
-    this.getDepartmentService().delete(departId);
-  }
-
-  @Override
-  @Deprecated
-  public List<WxCpDepart> departGet() throws WxErrorException {
-    return this.getDepartmentService().listAll();
   }
 
   @Override
