+++ weixin-java-cp/src/main/java/me/chanjar/weixin/cp/api/impl/WxCpUserServiceImpl.java
@@ -7,6 +7,7 @@ import me.chanjar.weixin.cp.api.WxCpService;
 import me.chanjar.weixin.cp.api.WxCpUserService;
 import me.chanjar.weixin.cp.bean.WxCpUser;
 import me.chanjar.weixin.cp.util.json.WxCpGsonBuilder;
+import org.apache.commons.lang3.StringUtils;
 
 import java.util.List;
 
@@ -112,4 +113,16 @@ public class WxCpUserServiceImpl implements WxCpUserService {
       );
   }
 
+  @Override
+  public int invite(String userId, String inviteTips) throws WxErrorException {
+    String url = "https://qyapi.weixin.qq.com/cgi-bin/invite/send";
+    JsonObject jsonObject = new JsonObject();
+    jsonObject.addProperty("userid", userId);
+    if (StringUtils.isNotEmpty(inviteTips)) {
+      jsonObject.addProperty("invite_tips", inviteTips);
+    }
+    String responseContent = this.mainService.post(url, jsonObject.toString());
+    JsonElement tmpJsonElement = new JsonParser().parse(responseContent);
+    return tmpJsonElement.getAsJsonObject().get("type").getAsInt();
+  }
 }
